fix(spec): one diagnostic prose source for dual error codes (#987)
Move PIPELINE_ERROR_CATALOG into @ggsvelte/spec, share summary/fix for
the 18 dual-channel codes via DUAL_ERROR_PROSE, and rename validation
scale-manual-domain-range to color-manual-domain-range. Core re-exports
the pipeline catalog; docs import pipeline error prose from spec.

Migration: https://ggsvelte.sh/guide/upgrading#0-11-to-0-12
