+++ lifecycle.json
@@ -1120,6 +1120,10 @@
           "kind": "type",
           "lifecycle": "experimental"
         },
+        "PIPELINE_ERROR_CATALOG": {
+          "kind": "value",
+          "lifecycle": "experimental"
+        },
         "POINT_SHAPE_NAMES": {
           "kind": "value",
           "lifecycle": "experimental"
@@ -1145,6 +1149,14 @@
           "lifecycle": "experimental"
         },
         "PathParams": {
+          "kind": "type",
+          "lifecycle": "experimental"
+        },
+        "PipelineErrorCatalogEntry": {
+          "kind": "type",
+          "lifecycle": "experimental"
+        },
+        "PipelineErrorCode": {
           "kind": "type",
           "lifecycle": "experimental"
         },
